requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-decouple==3.7
python-http-client==3.3.7
requests==2.28.2
six==1.11.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
urllib3==1.26.14
sendgrid==6.9.7
numerize==0.12
drf-api-logger==1.1.12
# openai==0.25.0
openai==0.27.2
BeautifulSoup4==4.12.0
celery==5.2.7
langchain==0.0.144
scikit-learn==1.2.2
nltk==3.8.1
django_celery_results==2.5.0
redis==4.5.4
django-storages
boto3
PyPDF2
spacy
langchain==0.0.308
openai==0.27.2
tiktoken==0.5.1
faiss-cpu==1.7.4